Split Helm upgrade execution and monitoring into separate classes
